move code to utils to simplify package dependencies
